requirements.txt
Flask==0.12.1
transitions==0.5.0
pygraphviz==1.3.1
python-telegram-bot==5.3.0
boto3==1.4.4
gTTS==1.2.0
beautifulsoup4==4.4.1

